Add license details before publishing.
This is the commit message #2: Escape numbered list starters so they don't and linkify links.
This is the commit message #3: Remove all the leading whitespace from the license texts - it causes github to format them as code blocks with copy icons. Add blank lines between list items to make them clearer to read.
This is the commit message #4: Those backslashes showed up as backslashes, out they go.
This is the commit message #5: Hopefully the last of the LICENSE commits - I feel a rebase -i coming on. I don't usually hide the sausage making but by golly this has gotten out of hand.
This is the commit message #6: WIP
This is the commit message #7: WIP
This is the commit message #8: Screw it, links for the CC licenses, enough time wasted already.
This is the commit message #9: FFS...